# openapi.yaml
openapi: "3.1.0"
info:
  title: Service-Controller-NotebookUm API
  version: "1.0"
  description: >
    Punto de entrada HTTP público del ecosistema NotebookUm.
    Implementa el patrón Strangler, validación en borde, Circuit Breaker,
    Bulkhead, Rate Limiting y Saga Orchestrator para el flujo de upload.
  contact:
    name: Equipo NotebookUm
servers:
  - url: http://controller.universidad.localhost/api/v1
    description: Desarrollo local (Traefik)

# ─────────────────────────────────────────
# Schemas reutilizables
# ─────────────────────────────────────────
components:
  schemas:
    ProblemDetail:
      description: RFC 9457 — Problem Details for HTTP APIs
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
          example: "https://notebookum.example/errors/file-too-large"
        title:
          type: string
          example: "File Too Large"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "The uploaded file exceeds the maximum size of 25MB."
        instance:
          type: string
          example: "/api/v1/documento/upload?requestId=abc123"
...
  headers:
    X-Correlation-ID:
      description: >
        Identificador único de la petición. Generado por el controller si el
        cliente no lo envía. Propagado a todos los upstreams.
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Falta o es inválido el header de autenticación.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/X-Correlation-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Forbidden:
      description: El usuario autenticado no tiene acceso al recurso.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/X-Correlation-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Recurso no encontrado.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/X-Correlation-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    ServiceUnavailable:
      description: >
        Upstream no disponible, circuit breaker abierto, o bulkhead saturado.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/X-Correlation-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Token de autenticación. El controller extrae el user_id del header
        Authorization y lo propaga como X-User-ID a los upstreams.

# ─────────────────────────────────────────
# Paths
# ─────────────────────────────────────────
paths:
...
  /documents/{document_id}:
    get:
      operationId: getDocument
      summary: Obtener metadatos y estado de un documento
      description: >
        Devuelve metadatos y estado actual del documento (PENDING/EXTRACTING/
        PROCESSING/READY/FAILED). Útil para polling del Saga.
        Tipo: query → service-persistence-read.
      tags: [documents]
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Metadatos del documento.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/X-Correlation-ID"
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [PENDING, EXTRACTING, PROCESSING, READY, FAILED]
                  filename:
                    type: string
                  owner_id:
                    type: string
                    format: uuid
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ── Resúmenes ─────────────────────────────────────────────────────────────
...